requirements.txt
whois21
dnspython
dnslib
redis
requests
